Change message length to uint32_t to fit camera data

+++ include/common/dprotocol/d_protocol.hpp
@@ -1,6 +1,7 @@
 #ifndef D_PROTOCOL_HPP
 #define D_PROTOCOL_HPP
 
+#include <iostream>
 #include "dprotocol/circular_byte_array.hpp"
 #include "dprotocol/d_protocol_message.hpp"
 
@@ -12,7 +13,7 @@ public:
   // preserved space
   uint16_t message_header;
   uint8_t protocol_version;
-  uint16_t messageLength;
+  uint32_t messageLength;
 
   uint8_t retrieveMessage(CircularByteArray *cba, DProtocolMessage *message);
